requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
discord.py==1.7.3
aiohttp==3.7.4.post0
beautifulsoup4==4.9.3
bs4==0.0.1
